♻️  refactor: type 정의 파일 생성 및 적용
