+++ src/domain/Restaurant.ts
@@ -1,20 +1,4 @@
-export type Category =
-  | "전체"
-  | "한식"
-  | "중식"
-  | "양식"
-  | "일식"
-  | "아시안"
-  | "기타";
-type Distance = 5 | 10 | 15 | 20 | 30;
-
-export interface RestaurantForm {
-  category: Category;
-  name: string;
-  distance: Distance;
-  description?: string;
-  link?: string;
-}
+import { RestaurantForm } from "../types";
 
 export class Restaurant {
   #information: RestaurantForm;
